Clean up CSS afterprint.
This fixes a few bugs that could occur when there were data URIs
and/or @imports in the original CSS (styles would be borked after
returning to the screen view). It also helps ensure that DOM
inspection is as WTF-less as possible.

A new `data-importeddatauri` attribute is used to keep track of CSS
state and the prefixed "raw" (unimported) CSS is stored in a property
named `beforeBeforePrintTextContent` on the HTMLStyleElement objects
themselves.