Ignore stylesheets without (data-)href attribute
In the case when stylesheet has not (data-)href attribute (for example when using Angular directive ng-href. For this case href attribute will not available while controller variables will not available) prefixfree should not post-process stylesheet because it will throw an error. For this case stylesheet post-processing logic should process DynamicDom plugin.